pyproject.toml
[project]
name = "speakflow"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.12.0",
    "edge-tts>=7.2.7",
    "flask>=3.1.2",
    "numba>=0.57.0",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "pydub>=0.25.1",
    "scipy>=1.15.3",
    "sounddevice>=0.5.3",
    "torch>=2.9.1",
    "transformers>=4.57.3",
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
